Checkpoint a functional HTTP01 Certificate implementation.
This doesn't have any unit test coverage (yet), it also doesn't have any configmap for controlling the ACME endpoint it speaks to (currently Let's Encrypt staging is checked in due to higher rate limits).

This commit also contains `./cmd/sample` which is code/config for standalone testing of the key libraries that the controller uses outside of a control loop.